requirements.txt
# PyTorch CPU Wheels
--extra-index-url https://download.pytorch.org/whl/cpu

# Python libs
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.pipeline==3.0.1
pyannote.metrics==3.2.1
pytorch-lightning==2.2.5
lightning==2.2.5
lightning-utilities==0.11.9
torchmetrics==1.3.2
pytorch-metric-learning==2.9.0
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1
noisereduce==3.0.0
pydub==0.25.1
soxr==1.0.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
speechbrain==0.5.16
asteroid-filterbanks==0.4.0
transformers==4.35.2
huggingface-hub==0.20.3
sentence-transformers==2.2.2
tokenizers==0.15.0
safetensors==0.4.1
numpy==1.24.3
scipy==1.10.1
numba==0.60.0
mpmath==1.3.0
sympy==1.14.0
scikit-learn==1.3.0
pandas==2.3.3
matplotlib==3.9.4
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
Pillow==11.3.0
praat-parselmouth==0.4.7
plotly==5.18.0
tabulate==0.9.0
fastapi==0.104.1
starlette==0.27.0
uvicorn==0.24.0
aiohttp==3.8.5
async-timeout==4.0.3
aiosignal==1.4.0
sniffio==1.3.1
websockets==15.0.1
h11==0.16.0
httptools==0.7.1
yarl==1.22.0
attrs==25.4.0
six==1.17.0
regex==2026.1.15
tokenizers==0.15.0
packaging==24.2
importlib_resources==6.5.2
platformdirs==4.4.0
pathlib==1.0.1
pytz==2025.2
typing_extensions==4.15.0
typing-inspection==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
tomli==2.4.0
zipp==3.23.0
click==8.1.8
typer==0.21.1
omegaconf==2.3.0
hydra-core==1.3.2
decorator==5.2.1
lazy_loader==0.4
shellingham==1.5.4
SQLAlchemy==2.0.46
alembic==1.16.5
propcache==0.4.1
filelock==3.19.1
pooch==1.8.2
PyYAML==6.0.1
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.15
Markdown-it-py==3.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==11.3.0
primePy==1.3
lazy_loader==0.4
joblib==1.5.3
msgpack==1.1.2
networkx==3.2.1
pydantic==2.12.5
pydantic_core==2.41.5
einops==0.8.2
Mako==1.3.10
mdurl==0.1.2
frozenlist==1.8.0
greenlet==3.2.4
